% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze the provided expression data for ubiquitously expressed genes.}
\usage{
analyze(data)
}
\arguments{
\item{data}{A \code{data.table} in normalized, long format. There should be a
\code{gene} column containing Ensembl gene IDs, a \code{sample} column containing
abitrary sample identifiers that are unique per sample and an \code{expression}
column containing the actual expression value for each given combination
of gene and sample.}
}
\value{
A \code{data.table} containing all computed values per gene.
}
\description{
Analyze the provided expression data for ubiquitously expressed genes.
}
