% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Create a shiny application for Ubigen.}
\usage{
app(custom_dataset = NULL)
}
\arguments{
\item{custom_dataset}{This allows to set a custom dataset (return value of
\code{\link[=analyze]{analyze()}}) as the default dataset of the UI.}
}
\description{
Create a shiny application for Ubigen.
}
\seealso{
\code{\link[=run_app]{run_app()}} for immediately running the application.
}
