% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{box_plot}
\alias{box_plot}
\title{Create a plot comparing some genes with the overall ranking.}
\usage{
box_plot(ranked_data, highlighted_genes)
}
\arguments{
\item{ranked_data}{The ranking to visualize.}

\item{highlighted_genes}{Genes that will be compared.}
}
\value{
A \code{plotly} figure.
}
\description{
Create a plot comparing some genes with the overall ranking.
}
