% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmap}
\alias{cmap}
\title{The results from the analysis on CMap data.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12382 rows and 17 columns.
}
\usage{
cmap
}
\description{
The results from the analysis on CMap data.
}
\keyword{datasets}
