% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{A \code{data.table} containig data on genes and their expression behavior.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 55507 rows and 20 columns.
}
\usage{
genes
}
\description{
A \code{data.table} containig data on genes and their expression behavior.
}
\keyword{datasets}
