% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_all}
\alias{gtex_all}
\title{The results from the analysis across all GTEx samples.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 55507 rows and 17 columns.
}
\usage{
gtex_all
}
\description{
The results from the analysis across all GTEx samples.
}
\keyword{datasets}
