% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_tissues}
\alias{gtex_tissues}
\title{The results from the analysis across GTEx tissue aggregated data.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 56156 rows and 17 columns.
}
\usage{
gtex_tissues
}
\description{
The results from the analysis across GTEx tissue aggregated data.
}
\keyword{datasets}
