% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_tissues}
\alias{hpa_tissues}
\title{The results from the analysis across Human Protein Atlas' tissue aggregated
data.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20090 rows and 17 columns.
}
\usage{
hpa_tissues
}
\description{
The results from the analysis across Human Protein Atlas' tissue aggregated
data.
}
\keyword{datasets}
