% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rankings_comparison_plot}
\alias{rankings_comparison_plot}
\title{Create a scatter plot for comparing two different rankings.}
\usage{
rankings_comparison_plot(
  ranking_x,
  ranking_y,
  label_x = "Ranking X",
  label_y = "Ranking Y",
  highlighted_genes = NULL,
  use_percentiles = FALSE
)
}
\arguments{
\item{ranking_x}{The ranking to be shown on the X-axis.}

\item{ranking_y}{The ranking to be shown on the Y-axis.}

\item{label_x}{Axis title for the X-axis.}

\item{label_y}{Axis title for the Y-axis.}

\item{highlighted_genes}{Gene IDs for genes that should be highlighted}

\item{use_percentiles}{Display percentiles instead of scores.}
}
\value{
A \code{plotly} figure for rendering.
}
\description{
Create a scatter plot for comparing two different rankings.
}
