% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_api}
\alias{run_api}
\title{Run the Ubigen API.}
\usage{
run_api(host = "127.0.0.1", port = 3465)
}
\arguments{
\item{host}{The hostname to serve the API on.}

\item{port}{The port to serve the API on.}
}
\description{
This requires the \code{plumber} package to be installed.
}
