% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the application server.}
\usage{
run_app(host = "127.0.0.1", port = 3464)
}
\arguments{
\item{host}{The hostname to serve the application on.}

\item{port}{The port to serve the application on.}
}
\description{
Run the application server.
}
