% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rank_genes}
\alias{rank_genes}
\title{Rank genes based on how ubiquitous they are.}
\usage{
rank_genes(
  data = ubigen::gtex_all,
  cross_sample_metric = "above_95",
  cross_sample_weight = 0.5,
  level_metric = "median_expression_normalized",
  level_weight = 0.25,
  variation_metric = "qcv_expression_normalized",
  variation_weight = -0.25
)
}
\arguments{
\item{data}{The input data to use. This should either be the result of a
previous call to this function or the return value of \code{\link[=analyze]{analyze()}}.}

\item{cross_sample_metric}{Name of the column that should be used as the
metric measuring the expression across samples.}

\item{cross_sample_weight}{Relative weight that should be assigned to the
cross sample metric.}

\item{level_metric}{Name of the column that should be used to represent
overall expression levels.}

\item{level_weight}{Relative weight that should be assigned to the level
metric.}

\item{variation_metric}{Name of the column that should be used as the metric
representing variation in expression.}

\item{variation_weight}{Relative weight that should be assigned to the
variation metric.}
}
\value{
A \code{data.table} with gene data as well as the scores, ranks and
percentiles for each gene.
}
\description{
This function will compute a weighted average across multiple metrics that
define how ubiquitous a gene is based on its expression across samples.
}
